added gen to run validation, and credits validation, updated runs counter for user and program stats, and average credits cost limiter

+++ dto/src/dto/DashboardDTO.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class DashboardDTO {
 
-    public int credits;
+    public long credits;
     public ArrayList<UserStatDTO> userStats;
     public ArrayList<ExecutionHistoryDTO> executionHistory;
     public ArrayList<SProgramViewStatsDTO> programStats;
