+++ dto/src/dto/UserStatDTO.java
@@ -4,7 +4,7 @@ public class UserStatDTO {
     public String username;
     public int num_uploaded_programs;
     public int num_uploaded_functions;
-    public int avail_credits;
-    public int credits_spent;
+    public long avail_credits;
+    public long credits_spent;
     public int num_runs;
 }
